+++ lib/hz_toast.dart
@@ -58,3 +58,4 @@ export 'src/hz_toast.dart';
 export 'src/hz_model.dart';
 export 'src/hz_toast_widget.dart';
 export 'src/hz_type_enum.dart';
+export 'src/hz_alignment_enum.dart';